requirements.txt
Django>=4.1.1,<4.2
djangorestframework>=3.14.0,<3.15
psycopg2>=2.8.6,<2.9
django-filter>=22.1,<22.2